+++ libraries/GCS_MAVLink/GCS_MAVLink.cpp
@@ -68,7 +68,7 @@ void GCS_MAVLINK::lock_channel(mavlink_channel_t _chan, bool lock)
     }
 }
 #if XBEE_TELEM==ENABLED
-void GCS_MAVLINK::xbee_set_targ_add(uint64_t* add_list, uint8_t lenth)
+void GCS_MAVLINK::xbee_set_targ_add(uint8_t* add_list, uint8_t lenth)
 {
     if(chan==MAVLINK_COMM_2)
         mavlink_comm_port[chan]->xbee_set_targ_add(add_list, lenth);
